Rename files according to the new name
