+++ tests/integration/targets/filter_data_structure_merging/tasks/run.yml
