+++ tests/integration/targets/filter_data_structure_merging/vars/main.yml
